Use correct var type to fix tpm attestation report

+++ vim25/types/types.go
@@ -21930,7 +21930,7 @@ type HostDigestInfo struct {
 	DynamicData
 
 	DigestMethod string `xml:"digestMethod"`
-	DigestValue  []byte `xml:"digestValue"`
+	DigestValue  []int `xml:"digestValue"`
 	ObjectName   string `xml:"objectName,omitempty"`
 }
 
@@ -26239,7 +26239,7 @@ func init() {
 type HostTpmEventDetails struct {
 	DynamicData
 
-	DataHash       []byte `xml:"dataHash"`
+	DataHash       []int `xml:"dataHash"`
 	DataHashMethod string `xml:"dataHashMethod,omitempty"`
 }
